\name{mycombine.fixed}
\alias{mycombine.fixed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Helper function
}
\description{
Helper function for combining outputs from replicated studies in fixed design tests.
}
\usage{
mycombine.fixed(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
Lists. Outputs from different replicated studies.
}
}
\value{
A list with two components combining the outputs from replicated studies.
}
\author{
Sandipan Pramanik and Valen E. Johnson
}