\name{mycombine.seq.onesample}
\alias{mycombine.seq.onesample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Helper function
}
\description{
Helper function for combining outputs from replicated studies in one-sample tests using Sequential Bayes Factor.
}
\usage{
mycombine.seq.onesample(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
Lists. Outputs from different replicated studies.
}
}
\value{
A list with three components combining the outputs from replicated studies.
}
\author{
Sandipan Pramanik and Valen E. Johnson
}