\name{mycombine.seq.twosample}
\alias{mycombine.seq.twosample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Helper function
}
\description{
Helper function for combining results in two-sample tests using Sequential Bayes Factor.
}
\usage{
mycombine.seq.twosample(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
Lists. Outputs from different replicated studies.
}
}
\value{
A list with four components combining the outputs from replicated studies.
}
\author{
Sandipan Pramanik and Valen E. Johnson
}