\name{NAP-package}
\alias{NAP-package}
\alias{NAP}
\docType{package}
\title{
\packageTitle{NAP}
}
\description{
\packageDescription{NAP}
}
\details{
\tabular{ll}{ Package: \tab NAP\cr Type: \tab Package\cr Version: \tab
1.1\cr Date: \tab 2022-1-6\cr License: GPL (>= 2)}
}
\author{
\packageAuthor{NAP}

Maintainer: \packageMaintainer{NAP}
}
\references{
Pramanik, S. and Johnson, V. (2022). \emph{Efficient Alternatives for Bayesian Hypothesis Tests in Psychology. Psychological Methods. Just accepted.}

Johnson, V. and Rossell, R. (2010). \emph{On the use of non-local prior densities in Bayesian hypothesis tests. Journal of the Royal Statistical Society: Series B, 72:143-170.}
\href{https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-9868.2009.00730.x}{[Article]}
}
